feat: 提供IAM所需openAPI #4259
