+++ src/backend/ci/core/project/biz-project/src/main/kotlin/com/tencent/devops/project/resources/ServiceMoveProjectResourceImpl.kt
@@ -1,0 +1,16 @@
+package com.tencent.devops.project.resources
+
+import com.tencent.devops.common.web.RestResource
+import com.tencent.devops.project.api.service.ServiceMoveProjectResource
+import com.tencent.devops.project.pojo.Result
+import com.tencent.devops.project.service.ProjectService
+import org.springframework.beans.factory.annotation.Autowired
+
+@RestResource
+class ServiceMoveProjectResourceImpl @Autowired constructor(
+    private val projectService: ProjectService
+): ServiceMoveProjectResource {
+    override fun relationIamProject(projectCode: String, relationId: String): Result<Boolean> {
+        return Result(projectService.relationIamProject(projectCode, relationId))
+    }
+}
